Fix bug in geometry iterator delete subroutine (#852)
TYPE: bug fix

DESCRIPTION OF CHANGES:
The main bug fix is provided by two changes:
- the `i2d` and `i3d` members of the `mpas_geom_iter` class are initialized to null pointers
- the implicit `mpas_geom_iter::final` procedure is changed to an explicit `delete` procedure that is called in `mpas_geom_iter_delete_f90` _before_ the linked list pointer to the `mpas_geom_iter` object is deallocated

Those fix the logical statements in `mpas_geom_iter_delete`.

Additional changes are made to increase the robustness of the `GeometryIterator` class.  For example, the `GeometryIterator::Geometry &` class member is added to ensure that the `Geometry` address is valid until the destructor is called.

ISSUE:
Fixes #850

LIST OF MODIFIED FILES:
M       src/mpasjedi/GeometryIterator/GeometryIterator.cc
M       src/mpasjedi/GeometryIterator/GeometryIterator.h
M       src/mpasjedi/GeometryIterator/iter.F90
M       src/mpasjedi/GeometryIterator/mpas_geom_iter_interface.F90
M       src/mpasjedi/GeometryIterator/mpas_geom_iter_mod.F90

TESTS CONDUCTED: ctests, cheyenne, gnu, debug build

+++ src/mpasjedi/GeometryIterator/iter.F90
@@ -6,6 +6,8 @@
 !> Geometry iterator
 module iter_mod
 
+use iso_c_binding
+
 use mpas_geom_mod, only: mpas_geom
 
 implicit none
@@ -102,7 +104,7 @@ contains
 subroutine iter_setup(self, geom, cellIndex, levIndex)
   class(iter),              intent(inout) :: self
   type(mpas_geom), pointer, intent(in)    :: geom !< Pointer to geometry
-  integer,                  intent(in)    :: cellIndex, levIndex  !< starting index
+  integer(c_int),           intent(in)    :: cellIndex, levIndex  !< starting index
 
   ! Associate geometry
   self%geom => geom
